feat : GH-22 Add Movies Api service and Repository        Add test for MoviesRepository        Closes #22
