+++ build-logic/convention/src/main/kotlin/me/jerryokafor/ihenkiri/Jacoco.kt
@@ -235,7 +235,7 @@ internal fun Project.configureJacoco(
 
             // Required for JDK 11 with the above
             // https://github.com/gradle/gradle/issues/5184#issuecomment-391982009
-            excludes = listOf("jdk.internal.*")
+            excludes = listOf("jdk.internal.*","*")
         }
     }
 }
