+++ gradle/libs.versions.toml
@@ -43,12 +43,12 @@ jacoco = "0.8.10"
 junit4 = "4.13.2"
 testing = "6.9.6"
 robolectric = "4.8"
-testCore = "1.4.0"
-testRunner = "1.5.1"
+testCore = "1.5.0"
+testRunner = "1.5.2"
 testRules = "1.5.0"
-testJunit = "1.1.4"
+testJunit = "1.1.5"
 testTruth = "1.5.0"
-testMonitor = "1.6.0"
+testMonitor = "1.6.1"
 espressoCore = "3.5.1"
 googleTruth = "1.1.4"
 mockk = "1.13.7"
